Move similar classes to separate packages
